feat: Lights are mapped to zones
The Zone class has mutator and accessor methods for the tuple of lights
that belong to the zone. A light is assigned to a zone if it is closer
to the zone than adjacent zones. To avoid repeated values, lights in
the list (for that edge) are set to None, as inconsistent gap sizes may
otherwise cause a light to be assigned to two zones.