fix: Zones no longer overlap
The makeZones() function has been changed such that if a zone shares
the same positional coordinates as an existing zone, the new zone is
not added to the zone list (and therefore is not returned) and the
new zone's lights are combined into the existing zone's tuple of lights.
To easily check which zones overlap, the equality method of the Zone
class has been overridden to return True if the positional coordinates
are the same and False if they are not.