feat: Zones color averages can be accessed
Each zone now has an object of the ImageHandler class. This new class
has methods to update the image and get the colour average of the
image. To avoid redundantly storing the coordinate values from the Zone
class, only the image width and height are stored as they are needed to
iterate through coordinate values in the image. The Zone class also has
two new methods: setImage() for cropping and storing the image in the
image handler and getColor() for accessing the average color of the
image. Additionally, a type "clr" has been defined as a tuple of three
integers to enhance code readability.