feat: Performance adaptation for color averaging
The ImageHandler class now takes a "performance value" to determine the
number of pixels skipped on each axis when taking pixel samples during
the averaging process. It accounts for a bias to sample the top-left by
changing the starting coordinates based on the number of pixels skipped.
The Zone class, and its occurrences in the main program have been
modified for the performance value to be passed to ImageHandler objects.