feat: Communication between host and client
On the host, a server listens for a connection requrest from the client.
The server accepts the request from the client and creates a connection.
The host sends data to the client indefinitely as the client replies
with receipts. Upon the connection terminating, the client will always
attempt to reconnect.