feat: Complete functional host program.
makeZones() has been moved to the "Zone.py" file where it has been given
new parameters to replace the constants defined in the old "main.py".
"main.py" has been completely remade such that it now contains the code
for connection management that was prototyped in "pcconnection.py". It
adds to this functionality by providing real zone and screen capture
data so the light output is now an approximation of the screen border.