feat: Functional home page.
The home page has working start and stop buttons that control the main
program process. The favourite presets list allows the user to change
the current preset. To handle the data in the presets file, the preset
manager class is instantiated and the resulting object is passes as an
argument into the constructor method of the home page.